Refactoctoring CV into asciidoc.
Moving cv into asciidoc so that it can be generated into the the multiple formats from a single source.

Adding support for auto-generating short and long versions of the CV in all formats [pdf,docx,md] from a single script.

There is a single adoc file for the CV containing conditional attributes for the different types. This may in the future support further A/B testing of the CV content and formats.